+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - If no events are passed to `Client.with_listener`/`Component.with_listener` it will now try to infer
   the event type(s) from the callback's type-hints.
 
+### Fixed
+- Slash command parent command tracking.
+
 ## [2.5.3a1] - 2022-07-04
 ### Added
 - Support for sending attachments with `Context.respond`, `CommandError` and